Refactor RegisterSystems back into ScanResults
